MAT-7926: Refactor common class to base qicore package.
